feat(pt): add ignore_time param for PeopleTracker
This feature allows the user to set the '$ignore_time' parameter in the
PeopleTracker per the People HTTP Specification of the Mixpanel REST
API. The user simply includes 'ignore_time' as True/False in their
properties dictionary and it gets converted to '$ignore_time' and
stripped from the properties dictionary. Included are tests and
documentation for this new functionality.